+++ connectors/connector-msp/src/main/scala/com/taskadapter/connector/msp/GTaskToMSP.scala
@@ -50,7 +50,8 @@ class GTaskToMSP(mspTask: Task, resourceManager: ResourceManager) {
     mspTask.setMilestone(false)
     // TODO eliminate the boolean argument. split this function in two.
     if (keepTaskId) { // Setting the old unique id, true only in "save external id" operation
-      mspTask.setUniqueID(gTask.getId)
+      // TODO TA3 this is probably obsolete and won't work with JIRA text IDs
+      mspTask.setUniqueID(gTask.getId.toInt)
     }
     gTask.getFields.asScala.foreach { x =>
       processField(x._1, x._2)
